final commit version 1.0
