+++ controllers/bookController.js
@@ -1,14 +1,17 @@
 const fetch = require("node-fetch");
 const Book = require("../models/bookModel");
+const capitalize = require("../functions/capitalize");
+const yearify = require("../functions/yearify");
 // book_index (all books), book_details (single book), book_create_get , book_create_post, book_delete
 
+////////////////////// Get all books ////////////////////////////
 const book_index = (req, res) => {
   Book.find()
-    .sort({ createdAt: -1 })
+    .sort({ createdAt: -1 }) //allows sorting by most recent at top
     .then((result) => {
       res.render(
-        "books/index",
-        { title: "All Books", books: result }
+        "books/index", //render the index view
+        { title: "All Books", books: result } //set books variable to the result. this books variable is now accessible in the index view
         // to view json in postman, uncomment below
         // (err, html) => res.send({ result })
       );
@@ -17,17 +20,16 @@ const book_index = (req, res) => {
       console.log(error);
     });
 };
-
+////////////////////// Get individual book details ////////////////////////////
 const book_details = async (req, res) => {
-  const id = req.params.id;
-  if (!id) return;
+  const id = req.params.id; //set id variable to the individual book id (id from database)
+  if (!id) return; //guard clause
 
   await Book.findById(id)
     .then((result) => {
       res.render(
-        "books/details",
+        "books/details", //render details view page and inject book data in to be used by the view file
         { book: result, title: "Book Details" }
-        // (err, html) => res.send({ result })
       );
     })
     .catch((err) => {
@@ -40,17 +42,20 @@ const book_create_get = (req, res) => {
   res.render("books/create", { title: "Add new book" });
 };
 
+////////////////////// Create new book ////////////////////////////
 const book_create_post = async (req, res) => {
   // console.log(req.body);
   if (!req.body.isbn) {
+    //guard clause if no isbn entered
     res.status(400).send({ message: "You must enter an ISBN number!" });
     return null;
   }
 
-  const book = new Book(req.body);
-  let id = String(book._id);
+  const book = new Book(req.body); //initialise new Book object based on Book schema
+  let id = String(book._id); //set the id variable to the individual book id, need to use string to convert from ObjectId to just Id
   let url = `/books/info/${id}`;
 
+  // Fetch image cover from OpenLibrary API
   const cover = await fetch(
     `https://covers.openlibrary.org/b/isbn/${req.body.isbn}-L.jpg`,
     {
@@ -64,27 +69,26 @@ const book_create_post = async (req, res) => {
     // .then(book.save())
     .catch((err) => console.log(err)); //HERE res.url gives the API link to the cover page now we need to integrate it into img in HTML
 
+  //save the cover image in the database
   book
     .save()
     .then(function (result) {
       let id = String(result._id);
       let url = `/books/info/${id}`;
       // console.log(url);
-      res.redirect(url);
+      res.redirect(url); //redirect to the newly created book url (details page of new book)
     })
     .catch((err) => console.log(err));
 };
+
+////////////////////// Fetch book data ////////////////////////////
 const book_info = async (req, res) => {
   try {
     const id = req.params.id;
     let url;
-    // console.log(id);
+    const isbn = await Book.findById(id).then((result) => result.isbn.trim()); //get the isbn of the book in question
 
-    const isbn = await Book.findById(id).then((result) => result.isbn.trim());
-
-    console.log(isbn);
-
-    url = `https://openlibrary.org/api/volumes/brief/isbn/${isbn}.json`;
+    url = `https://openlibrary.org/api/volumes/brief/isbn/${isbn}.json`; //inject the isbn into the url template literal to fetch data from the API
 
     const json = await fetch(url, {
       headers: {
@@ -95,51 +99,23 @@ const book_info = async (req, res) => {
       // parse the data with res.json
       .then((res) => res.json())
       // then return another promise as the data
-      .then((result) => (bookData = result))
-      .then(res.redirect(`/books/${id}`));
+      .then((result) => (bookData = result)) //set the result to bookData variable
+      .then(res.redirect(`/books/${id}`)); //redirect back to the individual book url
 
-    const capitalize = (string) => {
-      return string
-        .toLowerCase()
-        .split(" ")
-        .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-        .join(" ");
-    };
-
-    const yearify = (year) => {
-      if (year.includes(",")) {
-        return year.split(", ")[1];
-      }
-      if (year.includes("-")) {
-        return year.split("-")[0];
-      }
-      if (
-        year.startsWith("J") ||
-        year.startsWith("F") ||
-        year.startsWith("M") ||
-        year.startsWith("A") ||
-        year.startsWith("S") ||
-        year.startsWith("O") ||
-        year.startsWith("N") ||
-        year.startsWith("D")
-      ) {
-        return year.split(" ")[1];
-      } else {
-        return year;
-      }
-    };
-
+    // bookData returns a nested records object within an object within which the data of interest exists- destructure this as below:
     const bookInfo = bookData.records[Object.keys(bookData.records)[0]];
 
+    //pull out and assign relevant data to variables
     const title = capitalize(bookInfo.data.title);
     const [{ name: authorName }] = bookInfo.data.authors;
     const [{ url: authorUrl }] = bookInfo.data.authors;
-    // const publishedDate = Number(bookInfo.data.publish_date.split("-")[0]);
     const publishedDate = yearify(bookInfo.data.publish_date);
     const pages = bookInfo.data.number_of_pages;
     const [{ name: publishers }] = bookInfo.data.publishers;
     const bookUrl = bookInfo.data.url;
     const imageCovers = bookInfo.data.cover.large;
+
+    // Get genre of book from the array:
     let subjects;
 
     const genres = bookInfo.data.subjects.forEach(function (genre) {
@@ -150,6 +126,7 @@ const book_info = async (req, res) => {
       }
     });
 
+    // update the book with the data as found above
     await Book.findByIdAndUpdate(
       id,
       {
@@ -186,9 +163,7 @@ const book_delete = (req, res) => {
 
 const book_edit = async (req, res, next) => {
   try {
-    // console.log(req.body);
     const id = req.params.id;
-    // console.log(id);
     const book = req.body;
 
     const bookUpdate = await Book.findByIdAndUpdate(
@@ -204,42 +179,18 @@ const book_edit = async (req, res, next) => {
         runValidators: true,
       }
     )
-      // .then((result) => console.log(result))
       .then(res.redirect(`/books/${id}`))
       .catch((err) => console.log(err));
-    console.log(book);
-    // res.status(200).render("books/index.ejs", {
-    //   title: "Book details updated",
-    //   data: bookUpdate,
-    // });
-    // .then(res.json({ redirect: `/books/edit/${id}` }));
-
-    // .then((result) => res.render("/books/edit", { book: book }));
-
-    // console.log(req.body);
-    // if (!bookUpdate) {
-    //   res.status(404).json({ status: "Book not found!" });
-    // }
-
-    // book.save().then((result) => console.log(result));
-
-    // res.status(200).json({
-    //   status: "Book details successfully updated.",
-    //   redirect: `/books/${id}`,
-    //   data: {
-    //     data: bookUpdate,
-    //   },
-    // });
   } catch (err) {
     console.log(err);
   }
 };
 
+//export all controller functions to be accessible in bookRoutes
 module.exports = {
   book_index,
   book_details,
   book_create_post,
-  // book_image_cover,
   book_create_get,
   book_delete,
   book_edit,
