+++ functions/capitalize.js
@@ -1,0 +1,9 @@
+const capitalize = (string) => {
+  return string
+    .toLowerCase()
+    .split(" ")
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(" ");
+};
+
+module.exports = capitalize;
